pyproject.toml
[tool.poetry]
name = "st_mcp"
version = "0.1.0"
description = "MCP 기반 주식 분석 시스템"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
matplotlib = "^3.7.0"
yfinance = "^0.2.0"
mplcursors = "^0.5.0"
mplfinance = "^0.12.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 